+++ src/index.js
@@ -15,25 +15,8 @@ const syncCommand = require('./commands/sync');
 
 // Import utilities
 const { setDebugMode } = require('./utils/debug');
+const displayBanner = require('./utils/banner');
 
-// ASCII Art Banner
-const banner = `
- .d8888b.  888                        888                   d8b          
-d88P  Y88b 888                        888                   Y8P          
-888    888 888                        888                                
-888        888  .d88b.  888  888  .d88888  .d88b.  88888b.  888 888  888 
-888        888 d88""88b 888  888 d88" 888 d88""88b 888 "88b 888 \`Y8bd8P' 
-888    888 888 888  888 888  888 888  888 888  888 888  888 888   X88K   
-Y88b  d88P 888 Y88..88P Y88b 888 Y88b 888 Y88..88P 888  888 888 .d8""8b. 
- "Y8888P"  888  "Y88P"   "Y88888  "Y88888  "Y88P"  888  888 888 888  888 
-                                                                          
- v${pkg.version} - Cloudonix Voice AI Connector
-
- Home: https://cloudonix.com
- Documentation: https://developers.cloudonix.com
- Discord: https://discord.gg/etCGgNh9VV
- GitHub: https://github.com/cloudonix
-`;
 
 program
     .name('cx-vcc')
@@ -43,10 +26,9 @@ program
       // Set debug mode based on the global flag
       const options = thisCommand.opts();
       setDebugMode(options.debug || false);
+      displayBanner();
     });
 
-// Display the banner on each execution
-console.log(chalk.cyan(banner));
 
 // Configure command
 program